image-caption-explorer/Backend/requirements.txt
fastapi>=0.95.0
uvicorn
requests
python-dotenv
Pillow
aiohttp>=3.8.1
pydantic
transformers>=4.35.0
torch>=2.1.0
typing-extensions
certifi
beautifulsoup4
urllib3>=1.26.0